requirements.txt
asgiref==3.11.0
click==8.3.1
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
uvicorn==0.38.0
